/dConnection: added a `resetDomain` so a $cnxn object with its cached statements can be re-purposed on another domain. Saves lots of memory.
Tweaked the filter logic in /dConnection to reset the struct cache when
filters are cleared.

In /dView I switched from a path-based to a query-based approach. The
path-based worked because I've had MultiViews option set in the apache
config files, but I'm working to eliminate that. The path-based
approach won't work without that, but the key-based approach works
regardless.