include tree in calls to field/ref Defs
This allows dynamic results for fields and refs, which something like
dtVersioned will need (and dsVersioned currently uses). Specifically:

traits and base classes can call static methods on the most-descended
child (rather than on the static class, which in most cases would not
be correct)

caching is done not just by gname, but by the full hierarchy. Without
this, the first time we encountered dsVersion, we'd store the refs for
(maybe dsPost) and those would come back when we walked the hierarchy
for, say, dsPage. Now dsPost and dsPage have separate entries in the
cache for their parent dsVersioned results.